Move `Columns` class into own file
This is to avoid circular dependencies in subsequent commits.
